Use a separate babel config for the local-cli and the packager
