+++ local-cli/server/checkNodeVersion.js
@@ -8,17 +8,17 @@
  */
 'use strict';
 
-const chalk = require('chalk');
-const formatBanner = require('./formatBanner');
-const semver = require('semver');
+var chalk = require('chalk');
+var formatBanner = require('./formatBanner');
+var semver = require('semver');
 
 module.exports = function() {
   if (!semver.satisfies(process.version, '>=4')) {
-    const engine = semver.satisfies(process.version, '<1 >=4')
+    var engine = semver.satisfies(process.version, '<1')
       ? 'Node'
       : 'io.js';
 
-    const message = 'You are currently running ' + engine + ' ' +
+    var message = 'You are currently running ' + engine + ' ' +
       process.version + '.\n' +
       '\n' +
       'React Native runs on Node 4.0 or newer. There are several ways to ' +
