+++ local-cli/server/server.js
@@ -9,7 +9,6 @@
 'use strict';
 
 const chalk = require('chalk');
-const checkNodeVersion = require('./checkNodeVersion');
 const formatBanner = require('./formatBanner');
 const parseCommandLine = require('../util/parseCommandLine');
 const path = require('path');
@@ -87,8 +86,6 @@ function _server(argv, config, resolve, reject) {
       path.resolve(process.cwd(), dir)
     )
     : config.getAssetRoots();
-
-  checkNodeVersion();
 
   console.log(formatBanner(
     'Running packager on port ' + args.port + '.\n\n' +
