+++ package.json
@@ -131,7 +131,9 @@
     "art": "^0.10.0",
     "babel-core": "^6.6.4",
     "babel-plugin-external-helpers": "^6.5.0",
+    "babel-plugin-transform-flow-strip-types": "^6.6.5",
     "babel-polyfill": "^6.6.1",
+    "babel-preset-modern-node": "^1.1.0",
     "babel-preset-react-native": "~1.6.0",
     "babel-register": "^6.6.0",
     "babel-types": "^6.6.4",
