+++ packager/babelRegisterOnly.js
@@ -12,20 +12,17 @@ Array.prototype.values || require('core-js/fn/array/values');
 Object.entries || require('core-js/fn/object/entries');
 Object.values || require('core-js/fn/object/values');
 
-var fs = require('fs');
-var path = require('path');
-
 var _only = [];
-
-function readBabelRC() {
-  var rcpath = path.join(__dirname, 'react-packager', 'rn-babelrc.json');
-  var source = fs.readFileSync(rcpath).toString();
-  return JSON.parse(source);
-}
 
 module.exports = function(onlyList) {
   _only = _only.concat(onlyList);
-  var config = readBabelRC();
-  config.only = _only;
-  require('babel-register')(config);
+
+  require('babel-register')({
+    presets: ['modern-node/4.0'],
+    plugins: [
+      'transform-flow-strip-types',
+    ],
+    only: _only,
+    sourceMaps: 'inline',
+  });
 };
